GitLab CI testing with multiple sets of installed VCS
Also rolled in:
* Unify GitLab-CI and Dockerfile dependency install scripts
* Allow tests to fail gracefully when proper VCS is not installed.

+++ test/repotool-tags-hg.sh
@@ -1,6 +1,7 @@
 ## Test listing tags in a Mercurial repository
 
 command -v hg >/dev/null 2>&1 || { echo "    Skipped, hg missing."; exit 0; }
+command -v git >/dev/null 2>&1 || { echo "    Skipped, git missing."; exit 0; }
 
 trap "rm -rf /tmp/test-repo$$ /tmp/target$$ /tmp/out$$" 0 12 2 15
 
@@ -16,4 +17,3 @@ case $1 in
 esac
 	      
 #end
-
